+++ libs/xmlpack/xmlpackelement.h
@@ -61,10 +61,10 @@ public:
     void setNewValueDeviation(const QString&, bool &ok); // попытка установить новое отклонение значения
     void nextRandValue(); //обновить значение m_value с учетом rand_deviation, (если rand_deviation == 0, то значение не измениться)
 
-    void setIntValueByPath(const QList<quint8>&, qint64, bool&);
-    void setDoubleValueByPath(const QList<quint8>&, double, bool&);
-    qint64 getIntValueByPath(const QList<quint8>&, bool&);
-    double getDoubleValueByPath(const QList<quint8>&, bool&);
+    void setIntValueByPath(const QList<quint16>&, qint64, bool&);
+    void setDoubleValueByPath(const QList<quint16>&, double, bool&);
+    qint64 getIntValueByPath(const QList<quint16>&, bool&);
+    double getDoubleValueByPath(const QList<quint16>&, bool&);
 
 
     bool isNode() const; //элемент является секцией
@@ -111,7 +111,7 @@ protected:
     void retransformArrChild(int); //размножить найденный элемент-массив среди детей
     void transformTimeSpec();//сформировать структуру элементов для типа TimeSpec
     void nextTimeValue(); //обновить значение m_value типа timespec
-    LXMLPackElement* nodeByPath(const QList<quint8>&);
+    LXMLPackElement* nodeByPath(const QList<quint16>&);
 
 private:
     quint32 sectionSize(const LXMLPackElement*) const; //возвращает размер в байтах секции целиком
